Rename some methods, reimplement some methods, add some methods.
renderAll becomes update
writeLine becomes drawLine
writeRect becomes drawRectangle

drawLine (writeLine) is reimplemented as Breseman line algorithm
drawRectangle (writeRect) is reimplemented as drawing 4 lines

drawFilledRectangle is implemented as drawing multiple lines

drawCircle is implemeted as midpoint circle algorithm
drawFilledCircle is poorly implemented as drawing decreasing concentric circles - this doesn't really work right but will do for now

invertText is implemented to be able to control if text is printed in inverted colours or not (eg for highlighting)

write is largely reimplemented, newlines (or println()) will advance to next line.  Running off the end will cause all lines to shift up one and the bottom line to be cleared.

gotoXY no longer commands the LCD to go to the position, it's not necessary since we always use a framebuffer
update (renderAll) no longer seeks to 0,0